+++ src/main/resources/database/dataBase.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS hotel.Users
 
 CREATE TABLE IF NOT EXISTS hotel.Rooms
 (
-  roomID     SMALLSERIAL PRIMARY KEY,
+  roomID     BIGINT PRIMARY KEY,
   roomNumber SMALLINT NOT NULL UNIQUE,
   classID    VARCHAR(10) DEFAULT 'STANDARD' CHECK (classID IN ('ECONOMY', 'STANDARD', 'FAMILY', 'LUX')),
   capacity   VARCHAR(10) DEFAULT 'SINGLE' CHECK (capacity IN ('SINGLE', 'DOUBLE', 'TRIPLE', 'QUAD')),
@@ -33,8 +33,8 @@ CREATE TABLE IF NOT EXISTS hotel.Requests
 
 CREATE TABLE IF NOT EXISTS hotel.ReservedRooms
 (
-  reservedRoomID SMALLSERIAL PRIMARY KEY,
-  roomNumber     SMALLSERIAL REFERENCES hotel.Rooms (roomID),
+  reservedRoomID BIGSERIAL PRIMARY KEY,
+  roomNumber     BIGINT REFERENCES hotel.Rooms (roomID),
   requestID      BIGINT REFERENCES hotel.Requests (requestID)
 );
